feat: useCheckboxState 커스텀훅 생성
